+++ SDK/Tests/ServiceTests/AttachmentService/AttachmentServiceTests.swift
@@ -160,7 +160,7 @@ class AttachmentServiceTests: XCTestCase {
         waitForExpectations(timeout: 5)
     }
 
-    //Test image saved as jpeg with an invalid format as image
+    // Test image saved as jpeg with an invalid format as image
     func testUploadAttachmentInvalidImageData() {
         let attachment = FhirFactory.createImageAttachmentElement()
         let document = FhirFactory.createDocumentReferenceResource(with: [attachment])