+++ SDK/Tests/ServiceTests/FHIRService/ResourceWithAttachmentsTests/FHIRServiceTests+QuestionnaireResponse.swift
@@ -108,7 +108,7 @@ class FhirStu3ServiceQuestionnaireResponseTests: XCTestCase {
         questionnaireResponse.item?.first?.answer?.first?.item = [questionnaireResponseItem1Answer1Item1]
         let originalRecord = DecryptedRecordFactory.create(fhir: questionnaireResponse)
 
-        //We expect that the parameter of the uploadAttachments method pass the attachments without an Id
+        // We expect that the parameter of the uploadAttachments method pass the attachments without an Id
         let expectedAttachmentsWithoutId = questionnaireResponse.allAttachments!.compactMap {
             ($0.copy() as! Attachment) // swiftlint:disable:this force_cast
         }